fix bad with wrong human box

+++ model_image/ObjectDetectionModel.py
@@ -29,7 +29,7 @@ class YoloDetector(ObjectDetectionModel):
         res = detect(img, self.model)[0]["det"]
         boxes, confidence = np.array([]), np.array([])
         if len(res):
-            xyxy = res[:, :4].numpy().astype(np.uint8)
+            xyxy = res[:, :4].numpy().astype(np.uint16)
             confidence = res[:, 4].numpy()
             classes = res[:, 5].numpy().astype(np.uint8)
